+++ src/View/Player/Controller/GameSettings.java
@@ -133,7 +133,7 @@ public class GameSettings {
 		// get game params
 		txtBankrupt.getValueFactory().setValue((Integer) _mng.getParam(Param.BANKRUPTCY));
 		Logger.log("Aquired param money: "+_mng.getParam(Param.STARTING_CASH));
-		txtInitialSumOFMoney.getValueFactory().setValue((Integer) _mng.getParam(Param.STARTING_CASH));
+		txtInitialSumOFMoney.getValueFactory().setValue(((Double) _mng.getParam(Param.STARTING_CASH)).intValue());
 		txtNumOfRounds.getValueFactory().setValue((Integer) _mng.getParam(Param.MAX_ROUNDS));
 		txtPaymentRelaseFromJail.getValueFactory().setValue((Integer) _mng.getParam(Param.RELEASE_FROM_JAIL));
